+++ addons/yat/src/commands/builtin/Cls.cs
@@ -12,9 +12,9 @@ namespace YAT.Commands
 	)]
 	public sealed class Cls : ICommand
 	{
-		public CommandResult Execute(CommandArguments args)
+		public CommandResult Execute(CommandData data)
 		{
-			args.Terminal.Clear();
+			data.Terminal.Clear();
 
 			return CommandResult.Success;
 		}