+++ addons/yat/src/commands/builtin/Sys.cs
@@ -19,10 +19,10 @@ namespace YAT.Commands
 	[Threaded]
 	public sealed class Sys : ICommand
 	{
-		public CommandResult Execute(CommandArguments args)
+		public CommandResult Execute(CommandData data)
 		{
-			var program = (string)args.ConvertedArgs["-program"];
-			var command = (string)args.ConvertedArgs["command"];
+			var program = (string)data.Arguments["-program"];
+			var command = (string)data.Arguments["command"];
 			var commandName = command.Split(' ')[0];
 			var commandArgs = command[commandName.Length..].Trim() ?? string.Empty;