+++ addons/yat/src/types/CommandData.cs
@@ -4,13 +4,13 @@ using YAT.Scenes.BaseTerminal;
 
 namespace YAT.Interfaces
 {
-	public record CommandArguments(
+	public sealed record CommandData(
 		YAT Yat,
 		BaseTerminal Terminal,
 		ICommand Command,
-		string CommandName,
-		string[] Arguments,
-		Dictionary<string, object> ConvertedArgs,
+		string[] RawData,
+		Dictionary<string, object> Arguments,
+		Dictionary<string, object> Options,
 		CancellationToken? CancellationToken
 	);
 }