+++ CHANGELOG.md
@@ -16,10 +16,14 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 -   The man command displays information when the command being checked is threaded.
+-   Renamed CommandArguments to CommandData.
+-   Separated converted arguments & options in CommandData.
+-   Renamed Arguments to RawData in CommandData record.
 
 ### Removed
 
 -   LogHelper class (use Log instead).
+-   CommandName variable from CommandData.
 
 ## [1.17.0-beta 2024-01-12]